Add dashboard to link to both projects
This displays links to Add and Find (which will be in a separate repo,
so we'll add links to the live version hosted on Heroku when it's
ready).

It also displays a link to the Book and Manage project (which is housed
in this repo, now nested under the `/book-and-manage` directory.