Turn multiple <select>s into autocomplete token fields
We got this idea from the job role / organisation inputs on the Civil
Service Jobs site [1]. That site uses the Chosen jQuery plugin [2], but
that plugin’s now been deprecated by the vendor, so I picked (fairly
randomly) another library that allows us to do this.

There are almost certainly a lot of accessibility questions about using
this library, and I only feel comfortable throwing this library in
because it’s a prototype.

The SlimSelect library comes with its own CSS, with no obvious
customisation points, so we need to hackily override some of it.

I haven’t yet figured out the best way to bundle external libraries and
their CSS - I’m not sure if / how NPM fits into the prototype build
system - so I’m just linking to the CDN files at the moment.

[1] https://www.civilservicejobs.service.gov.uk/csr/index.cgi
[2] https://harvesthq.github.io/chosen/