Add search for service user from Community API
This allows users to search for an existing service user by CRN (an
existing user in the local data set has a CRN of X320741) and view that
user's personal information. Unfortunately not all of the data we expect
from the Community API is present on this example service user, but this
will have to do until we get access to the VPN when we'll have access to
a more complete dataset.

Running this search requires running a local version of the [Community
API](https://github.com/ministryofjustice/community-api). You can run
this (and the [HMPPS
Auth](https://github.com/ministryofjustice/hmpps-auth) service required
for generating an authentication token) from that project by running
`docker-compose up`.

Initially I tried to use an auto-generated client using the
`swagger-codegen` cli. However, I believe the authentication for the
`/secure` endpoints on the `community-api` project haven't been
documented in the same way as they had been for the
`offender-assessments-api` (the generated documentation stated that
these routes weren't authenticated). After fighting with swagger for a
while, I opted to write my own rough version of a client, which won't
win any awards for code quality or error handling, but will hopefully do for
the purposes of demoing this work.