Service provider case worker can manage a referral
This introduces the _end_ of the case worker journey. It focuses on a
single referral, which has already had its initial appointment
scheduled. I started here because it seems like the gnarliest part of
the journey, logic-wise.

There are fast-forward links to allow a user research session to skip
past points in the journey that would require interaction from a third
party, or to skip repetitive actions.

This is based on the bottom half of the “Receive a referral caseworker”
Figma board [1].

I haven’t yet built the header which contains details of the referral
and the service user.

This introduces Moment as a dependency, for manipulating and displaying
dates.

I’ve made some assumptions about the meaning of the various statuses
that are displayed in the journeys. No doubt the correct meanings will
crystallize over time.

[1] https://www.figma.com/file/uHcKdueAgFtYyeAURhBuNw/Receive-a-referral-caseworker