Add save buttons to Referral screen
These will be disabled unless they have text in them. This
implementation is a bit hacky, but it should do for now. We may want to
change the `onblur` handler to an `onkeyup` (which will be triggered any
time a key is pressed) depending on testing, but this felt like we would
be calling the function too many times. We could also poll this with a
timeout if need be.