Split prototype versions by sprint
As a team we’ve agreed that we’d like our alpha output to include all of
the prototypes that we tested. This means that we’d like to be able to
look at the prototype from, for example, sprint 2, and then look at the
iterated version from sprint 5. This has precedent in other alpha
prototypes we’ve seen, such as DfE Hubs [1].

This splits the previous sprints’ prototypes into /sprint-2 and
/sprint-3 routes. It also copies the /sprint-3 content to /sprint-4,
because we know we’ll be iterating on it this sprint.

[1] https://github.com/dxw/dfe-hubs-prototype