Create caseworker intervention progress page
We move the creation of the action plan to a separate page, and repeat
the sessions list on the intervention progress page as the place for
starting a post-session assessment.

Action plan page still needs referral details to be moved to sidebar to
match designs.

Still need to build statuses and conditional template logic.

This is based on the Figma [1].

[1] https://www.figma.com/file/29GGWjrJwNrv8TvvZOQkVI/Receive-a-referral-caseworker-2nd-iteration