Add caseworker intervention page referral details
These templates are copied from manager/includes/ and
manager/intervention.html, just changed to accommodate the slightly
different data structure and naming in the caseworker journey. The
designs for the caseworker journey didn’t include the expanded sections,
so I’ve just assumed it’s the same as for the manager.

The new data (risks and needs) is copied from the manager journey’s static data.