Add initial pages of caseworker journey
These are stolen from the corresponding manager pages. The dashboard
content is based on the Figma [1]. I haven’t added the email page
because although it exists in the Figma board, it doesn’t appear to be
linked into the prototype that’s being used for the user research
sessions.

[1] https://www.figma.com/file/29GGWjrJwNrv8TvvZOQkVI/Receive-a-referral-caseworker-2nd-iteration