Allow users to assign other interventions
Users can now click on and assign different interventions and will be
taken to those specific interventions (rather than defaulting to the
first one as we were doing previously)

This is a large commit, but modifies the shape of the session data
defaults and manager static data so required changing a few sections of
the prototype to access this data in the correct way.

The data is still very duplicated at the moment, so I'd like to
introduce more variety in future.

Now that we can assign a few interventions, we can move "assigned"
interventions into the respective tab on the referrals index page.