Show a more realistic caseworker name in top banner
This means that a research participant who is shown the caseworker
journey after being shown the caseworker journey will see a caseworker
name that they saw in the select box on the first journey.

I’ve switched the caseworker order so that Liam gets shown - I think
this is the name the designers are familiar with (Ming requested Liam or
Jenny).