Make Alex River data consistent with "Refer"
We're going to be testing these 2 journeys with Probation Practitioners,
so it makes sense to line up the data between the 2 journeys a bit more.

Unfortunately this doesn't use any data entered by users yet, but we
might be able to use that later on in the journey when displaying, based
on what's in the session data.