Remove nested {% if … %}s from intervention page
We now build up the contents of the screen in a linear fashion - the
conditional logic is done by passing flags which are set in the route
(also making this logic easier to write since we’re not grappling with
Nunjucks template syntax).

This will allow us to start dynamically re-ordering the list.