Caseworker sees PP’s judgement on absences
In the real journey, after a caseworker marks a session as a no-show in
Record, the probation practitioner will see this in Monitor, and they’ll
be able to issue a judgement on whether it’s an acceptable absence. This
judgement will then be displayed on the sessions progress table.

Since Monitor and Record aren’t joined in the prototypes, we instead
provide fast-forward links to simulate the different PP judgements.

This is based on the design in Miro [1].

[1] https://miro.com/app/board/o9J_krWjKeQ=/?moveToWidget=3074457350770503488&cot=12