Have a single PP judgement fast-forward link
George G thought that it would be sufficient to have one link that
allows him to exhibit how the “acceptable” and “unacceptable” absences
are displayed to the caseworker.

Once a session has been assessed as a no-show, this link appears, and
adds an “acceptable” judgement to that session, and then assesses the
next session as a no-show, adding an “unacceptable” judgement (adding a
new session if necessary).

I’ll need to explain this to George and check it works for him.