Only show intervention tasks after Find done
This breaks the ordering of the initial list (numbers skip from 4 to 7),
which we’ll address next.

Worth noting that this does not pay any attention to _which_
interventions were selected in Find. The referral task list will always
show accommodation and social inclusion. Researchers will need to guide
participants to choose these in Find.