Add fast forward links for Monitor
Until now, we’ve been demoing the different states in Monitor by showing
different service users. For our end of alpha demo we’d like to show the
entire journey for a single service user. So now we add fast-forward
links which allow us to move a service user between all of the states
that can be represented in the prototype.