Structure the action plan around desired outcomes
This follows a conversation with the Dynamic Framework policy team. The
action plan should be structured around the desired outcomes which the
probation practitioner selected in the referral. The caseworker can then
attach “activities” to each of these desired outcomes, explaining the
actions they’ll take to reach the outcome.