Remove unqualified interventions from referrals
Before this change, a referral's interventions were all grouped together
in the "stages" table (even if only one intervention matched the stage,
the other ones would be shown in that stage due to incorrect logic)

This manually sets the interventions on a new object, removing those
that don't qualify for the stage of an intervention.

This is currently pretty gnarly logic, but it felt quicker than changing
the underlying data when time is quite short. I originally wrote it as a
reduce to avoid `filter`ing and `mapping` but this looked even less
readable