Determine stage based on `assignedCaseworker`
rather than `monitor` data - the `Receive` journey actually affects this
data, and some of the monitor states weren't lining up correctly.

There's a lot more tidying up we could do here but because we're doing
this for the Show & Tell tomorrow, I'm going to do the minimum.

Also change assigned Caseworker to Liam Johnson for fast forwarded links
for demo